Closes #3: Allow extension of hadoop nodes in the cluster definition data structure
Adding `:extends [spec spec ...]` to the `:groups` section in the hadoop
cluster definition data structure extends the configuration of the
nodes with the supplied specs. This extension is done last after all
the other hadoop-related extensions. E.g.:

```clojure
(def our-hadoop-cluster
  (h/hadoop-cluster
   {:cluster-prefix "hi-hadoop"
    :groups {:nn {:node-spec our-node
                  :count 1
                  :extends [common/with-tags common/with-mosh]
                  :roles #{:namenode :jobtracker}}
             :slave {:node-spec our-node
                     :count 1
                     :extends [common/with-tags]
                     :roles #{:datanode :tasktracker}}}
    :hadoop-config
     {:io.file.buffer.size 65536
      :config {
       :namenode {:jmx-port 3000}
       :secondary-namenode {:jmx-port 3001}
       :jobtracker {:jmx-port 3002}
       :datanode {:jmx-port 3003}
       :tasktracker {:jmx-port 3004}}}}))
```

should extend the `:nn` nodes with the specs: `common/with-tags` and
`common/with-mosh` and the `:slave` nodes with `common/with-tags`.