Update job_spec.clj
File permissions necessary if using local-file else it's set to the local user launching the job

I was launching the job on my machine as user. The local jar got transfered to the remote cluster under permissions user:user.
Hadoop got a permission denied when trying to access it. Setting the perms to hadoop:hadoop fixes this.